see descript
now you end node is a public boolean feild that can be toggled changing
the display and later behavure.
also added start connector and the on and off bit connectors. later will
add the logic that enables only the start node to have a start connector
coming from it. also need to enforce each node has zero or one on/off
bit connectors.
added the start node to the view on applicaion open